refactor/commit-traversal (#7)
Simplify the traversal of commits and make it more efficient.

Previously, I iterated every single commit, even if most of them were skipped using the `CommitIter.ForEach` func. Now, we only loop if the condition is satisfied with our since value, this can technically still be every single commit with a large value, but is much better for general use on larger repositories.

It's rather anecdotal, but I tried it myself before on a repository like kubernetes and it seemed as though my computer froze. Now it is much faster, since the iteration stops short.

Also added various commit tests, which are useful for any type of refactoring later.